fix cooldown and energy types

+++ clans/src/main/java/me/mykindos/betterpvp/clans/champions/skills/types/CooldownSkill.java
@@ -2,7 +2,7 @@ package me.mykindos.betterpvp.clans.champions.skills.types;
 
 public interface CooldownSkill {
 
-    int getCooldown(int level);
+    double getCooldown(int level);
 
     default boolean showCooldownFinished() {
         return true;
